requirements.txt
streamlit==1.32.0
matplotlib==3.8.2
numpy==1.26.3
wfdb==4.1.2
opencv-python==4.8.1.78
torch==2.1.2
torchvision==0.16.2
timm==0.9.12
pytorch-grad-cam==0.4.1
Pillow==10.1.0
scipy==1.11.4
neurokit2==0.2.7
seaborn==0.13.1
scikit-learn==1.3.2
pandas==2.1.4
fpdf==1.7.2
ffmpeg-python==0.2.0